Add files via upload
This file features an advanced machine learning project focused on improving the detection and diagnosis of thyroid diseases. By utilizing sophisticated data analysis and predictive modeling techniques, the project aims to deliver a reliable tool for the early and precise identification of thyroid disorders, which is crucial for effective treatment and management.